+++ src/commands/mod.rs
@@ -1,16 +1,23 @@
 pub mod about;
 pub mod clear;
 pub mod cooldown;
+pub mod cooldowns;
 pub mod disabled;
+pub mod drop;
 pub mod drops;
 pub mod emoji;
+pub mod emojis;
 pub mod gift;
 pub mod given;
+pub mod guild;
 pub mod guilds;
 pub mod info;
 pub mod levelup;
-pub mod moderate;
+pub mod levelups;
+pub mod moderation;
+pub mod moderations;
 pub mod module;
+pub mod modules;
 #[cfg(feature = "nlp-model")]
 pub mod mood;
 #[cfg(feature = "nlp-model")]
@@ -18,7 +25,9 @@ pub mod oracle;
 pub mod ping;
 pub mod rank;
 pub mod reactionrole;
+pub mod reactionroles;
 pub mod reminder;
+pub mod reminders;
 pub mod say;
 pub mod score;
 pub mod sql;
