+++ src/config.rs
@@ -56,23 +56,32 @@ pub enum CommandType {
     About,
     Info,
     Module,
+    Modules,
     Ping,
     Guild,
+    Guilds,
     Say,
     Sql,
     Clear,
     Reminder,
+    Reminders,
     Cooldown,
+    Cooldowns,
+    Drop,
     Drops,
     Emoji,
+    Emojis,
     Gift,
     Given,
     LevelUp,
-    Moderate,
+    LevelUps,
+    Moderation,
+    Moderations,
     Rank,
     Score,
     Top,
     ReactionRole,
+    ReactionRoles,
     Mood,
     Oracle,
     Tldr,
@@ -126,11 +135,12 @@ pub enum Channel {
     Voice,
     Category,
     News,
-    Store,
     NewsThread,
     PublicThread,
     PrivateThread,
     Stage,
+    Directory,
+    Forum,
     Unknown,
 }
 
@@ -201,11 +211,12 @@ impl Into<ChannelType> for Channel {
             Channel::Voice => ChannelType::Voice,
             Channel::Category => ChannelType::Category,
             Channel::News => ChannelType::News,
-            Channel::Store => ChannelType::Store,
             Channel::NewsThread => ChannelType::NewsThread,
             Channel::PublicThread => ChannelType::PublicThread,
             Channel::PrivateThread => ChannelType::PrivateThread,
             Channel::Stage => ChannelType::Stage,
+            Channel::Directory => ChannelType::Directory,
+            Channel::Forum => ChannelType::Forum,
             Channel::Unknown => ChannelType::Unknown,
         }
     }
