+++ src/events/ready.rs
@@ -69,7 +69,7 @@ async fn setup_commands(ctx: &Context, rdy: Ready) {
     let commands = create_global_commands(ctx, &config).await;
     // Set permissions of the global commands per guild
     for guild in rdy.guilds {
-        add_permissions(ctx, &config, guild.id(), &commands).await;
+        add_permissions(ctx, &config, guild.id, &commands).await;
     }
     info!("{}", INFO_CMD_GLOBAL);
